Working on getting all projectiles to function. They were reading in mass as an Int, when mass is a double. This could be happening in other places in the code. Mass is a double!

+++ src/raven/game/armory/Bolt.java
@@ -26,7 +26,7 @@ public class Bolt extends RavenProjectile {
 				RavenScript.getInt("Bolt_Damage"),
 				RavenScript.getDouble("Bolt_Scale"),
 				RavenScript.getDouble("Bolt_MaxSpeed"),
-				RavenScript.getInt("Bolt_Mass"),
+				RavenScript.getDouble("Bolt_Mass"),
 				RavenScript.getDouble("Bolt_MaxForce")
 		);	
 	}
